+++ packages/iov-core/src/jsrpcsigningserver.spec.ts
@@ -1,3 +1,4 @@
+// tslint:disable: deprecation
 import {
   Address,
   Algorithm,
@@ -189,7 +190,7 @@ describe("JsRpcSigningServer", () => {
     bnsConnection.disconnect();
   });
 
-  it("send a signing request to service", async () => {
+  it("handles signing requests", async () => {
     pendingWithoutBnsd();
     pendingWithoutEthereum();
 
@@ -241,7 +242,7 @@ describe("JsRpcSigningServer", () => {
 
     const result = await firstEvent(bnsConnection.liveTx({ id: transactionId }));
     if (!isConfirmedTransaction(result)) {
-      throw new Error("Confirmed transaction extected");
+      throw new Error("Expected confirmed transaction");
     }
     expect(result.transactionId).toEqual(transactionId);
     expect(result.transaction).toEqual(send);